+++ CHANGELOG.md
@@ -2,9 +2,10 @@
 
 ## Unreleased
 
-## Features
+### Features
 
 - Update normalization of metrics keys, tags and values ([#3332](https://github.com/getsentry/sentry-java/pull/3332))
+- Enable backpressure management by default ([#3284](https://github.com/getsentry/sentry-java/pull/3284))
 
 ### Fixes